+++ src/app/(app)/cabinet/page.tsx
@@ -1,4 +1,4 @@
-import HomeWrapper from "@/app/_components/HomeWrapper";
+import HomeWrapper from "@/app/_components/oldHomeWrapper";
 
 export default async function Home() {
   return <HomeWrapper key="cabinet" type="cabinet" />;
