+++ src/app/(app)/fridge/page.tsx
@@ -1,4 +1,4 @@
-import HomeWrapper from "@/app/_components/HomeWrapper";
+import HomeWrapper from "@/app/_components/oldHomeWrapper";
 
 export default async function FridgePage() {
   return <HomeWrapper key="fridge" type="fridge" />;
